Random Generator method -> to assign different repositories
