+++ src/Info-Pages/Welcome.js
@@ -2,11 +2,12 @@ import React from "react";
 import "./InfoPages.css";
 import { useNavigate } from "react-router-dom";
 import { useState } from "react";
-import { Button, Checkbox, Text } from "@chakra-ui/react";
+import { Button, Checkbox, Text, Spinner } from "@chakra-ui/react";
 
 export default function Welcome() {
   const [checked, setChecked] = useState(false);
   const [disabledButton, setDisabled] = useState(true);
+  const [isSubmitting, setIsSubmitting] = useState(false);
 
   const navigate = useNavigate();
 
@@ -20,6 +21,28 @@ export default function Welcome() {
       setDisabled(false);
     }
   }
+
+  const assignARandomRepository = () => {
+    setIsSubmitting(true);
+    document
+      .querySelector("#submit-button")
+      .setAttribute("disabled", isSubmitting);
+
+    const rndInt = Math.floor(Math.random() * 4) + 1;
+    console.log(rndInt);
+    if (rndInt === 1) {
+      window.location.href = "https://e-voting-study-1.netlify.app/info-1";
+    }
+    if (rndInt === 2) {
+      window.location.href = "https://e-voting-study-2.netlify.app/info-1";
+    }
+    if (rndInt === 3) {
+      window.location.href = "https://e-voting-study-3.netlify.app/info-1";
+    }
+    if (rndInt === 4) {
+      window.location.href = "https://e-voting-study-4.netlify.app/info-1";
+    }
+  };
 
   return (
     <div className="container-info-pages">
@@ -35,10 +58,12 @@ export default function Welcome() {
           I agree.
         </Checkbox>
         <Button
-          onClick={() => navigate("/info-1")}
+          onClick={() => assignARandomRepository()}
           className="red-btn"
           disabled={disabledButton}
+          id="submit-button"
         >
+          {isSubmitting && <Spinner size="sm" mr={"1rem"} />}
           Start
         </Button>
       </div>
